+++ R/growthModelSim.R
@@ -1,29 +1,28 @@
-#'Dynamics plots to explore typical fisheries growth models.
+#' @title Dynamics plots to explore typical fisheries growth models.
 #'
-#'Plots hypothetical size-at-age for one of seven possible parameterizations of the von Bertalanffy, three possible parameterizations of the Gompertz, and the Schnute growth models.  Slider bars are used to dynamically alter the parameters of each model.
+#' @description Plots hypothetical size-at-age for one of seven possible parameterizations of the von Bertalanffy, three possible parameterizations of the Gompertz, and the Schnute growth models.  Slider bars are used to dynamically alter the parameters of each model.
 #'
-#'This function can be used to explore the \dQuote{shape} of the growth models for various choices of the parameters.  In this usage the \code{x} and \code{y} arguments should be (left) set at \code{NULL}.  This function can also be used to visually \dQuote{fit} a growth model to a set of observed lengths and ages.  This usage may be used to provide reasonable starting values for the parameters when fitting the growth model to the data with non-linear least-squares.  The observed data are plotted by including a formula of the form \code{length~age} in \code{x} and a data frame from which to draw the variables in the formula in the \code{data} arguments.
+#' @details This function can be used to explore the \dQuote{shape} of the growth models for various choices of the parameters.  In this usage the \code{x} and \code{y} arguments should be (left) set at \code{NULL}.  This function can also be used to visually \dQuote{fit} a growth model to a set of observed lengths and ages.  This usage may be used to provide reasonable starting values for the parameters when fitting the growth model to the data with non-linear least-squares.  The observed data are plotted by including a formula of the form \code{length~age} in \code{x} and a data frame from which to draw the variables in the formula in the \code{data} arguments.
 #'
-#'The \code{type} argument is used to choose a type of growth model and must be one of the following (the models can be seen with \code{\link{growthModels}} and \code{\link{vbModels}}):
+# 'The \code{type} argument is used to choose a type of growth model and must be one of the following (the models can be seen with \code{\link{growthModels}} and \code{\link{vbModels}}):
 #'
-#'\tabular{ll}{
-#'\code{"vbTypical"} \tab The "typical" Beverton-Holt parameterized von Bertalanffy model.\cr
-#'\code{"vbOriginal"} \tab The original parameterization from von Bertalanffy.\cr
-#'\code{"vbMooij"} \tab The Mooij et al (1999) paramaterization of the von Bertalanffy model.\cr
-#'\code{"vbGalucciQuinn"} \tab The Galucci & Quinn (1979) parameterization of the von Bertalanffy model.\cr 
-#'\code{"vbSchnute"} \tab The Schnute-like paramaterization of the von Bertalanffy model.\cr 
-#'\code{"vbTypicalW"} \tab The "typical" Beverton-Holt parameterized von Bertalanffy model, but for weights rather than lengths (thus, includes one more parameter).\cr
-#'\code{"vbOriginalW"} \tab The original parameterization from von Bertalanffy, but for weights rather than lengths (thus, includes one more parameter).\cr
-#'\code{"Gompertz1"} \tab The "first" parameterization of the Gompertz model.\cr 
-#'\code{"Gompertz2"} \tab The "second" parameterization of the Gompertz model.\cr 
-#'\code{"Gompertz3"} \tab The "third" parameterization of the Gompertz model.\cr
-#'\code{"schnute"} \tab The Schnute(1981) four-parameter general growth model.
-#'}
+#' \tabular{ll}{
+#' \code{"vbTypical"} \tab The "typical" Beverton-Holt parameterized von Bertalanffy model.\cr
+#' \code{"vbOriginal"} \tab The original parameterization from von Bertalanffy.\cr
+#' \code{"vbMooij"} \tab The Mooij et al (1999) paramaterization of the von Bertalanffy model.\cr
+#' \code{"vbGallucciQuinn"} \tab The Gallucci & Quinn (1979) parameterization of the von Bertalanffy model.\cr 
+#' \code{"vbSchnute"} \tab The Schnute-like paramaterization of the von Bertalanffy model.\cr 
+#' \code{"vbTypicalW"} \tab The "typical" Beverton-Holt parameterized von Bertalanffy model, but for weights rather than lengths (thus, includes one more parameter).\cr
+#' \code{"vbOriginalW"} \tab The original parameterization from von Bertalanffy, but for weights rather than lengths (thus, includes one more parameter).\cr
+#' \code{"Gompertz1"} \tab The "first" parameterization of the Gompertz model.\cr 
+#' \code{"Gompertz2"} \tab The "second" parameterization of the Gompertz model.\cr 
+#' \code{"Gompertz3"} \tab The "third" parameterization of the Gompertz model.\cr
+#' \code{"schnute"} \tab The Schnute(1981) four-parameter general growth model.
+#' }
 #'
-#' @param type A single character string that indicates which growth model to use.  See details.
-#' @param x An optional vector that contains observed ages or a formula.  See details.
-#' @param y An optional vector that contains observed lengths.  See details.
+#' @param formula An optional formula of the form \code{len~age}.  See details.
 #' @param data A data frame from which the variables in the formula should be drawn.
+#' @param type A single character string that indicates which growth model to use.  See details.
 #' @param max.len A single numeric that indicates the maximum length to use in the simulations.
 #' @param max.wt A single numeric that indicates the maximum weight to use in the simulations (only used of \code{type=} \code{"vbTypicalW"} or \code{"vbOriginalW"}.
 #'
@@ -35,7 +34,7 @@
 #'
 #' @references Francis, R.I.C.C.  1988.  Are growth parameters estimated from tagging and age-length data comparable?  Canadian Journal of Fisheries and Aquatic Sciences, 45:936-942.
 #'
-#'Galucci, V.F. and T.J. Quinn II. 1979.  Reparameterizing, fitting, and testing a simple growth model.  Transactions of the American Fisheries Society, 108:14-25.
+#'Gallucci, V.F. and T.J. Quinn II. 1979.  Reparameterizing, fitting, and testing a simple growth model.  Transactions of the American Fisheries Society, 108:14-25.
 #'
 #'Mooij, W.M., J.M. Van Rooij, and S. Wijnhoven.  1999.  Analysis and comparison of fish growth from small samples of length-at-age data: Detection of sequal dimorphism in Eurasian perch as an example.  Transactions of the American Fisheries Society 128:483-490.
 #'
@@ -59,96 +58,23 @@
 #' # get Smallmouth Bass data from FSA package
 #' data(SMBassWB)
 #'
+#' # interactively "fit" the typical paramaterization of the von Bertalanffy model to the data
+#' growthModelSim(lencap~agecap,data=SMBassWB)
+#'
 #' # interactively "fit" the second paramaterization of the Gompertz model to the data
-#' growthModelSim(type="Gompertz2",lencap~agecap,data=SMBassWB,max.len=500)
+#' growthModelSim(lencap~agecap,data=SMBassWB,type="Gompertz2")
 #'
 #' } ## END IF INTERACTIVE MODE
 #'
-#' @export
-growthModelSim <- function(type=c("vbTypical","vbOriginal","vbGalucciQuinn","vbMooij",
-                                  "vbSchnute","vbTypicalW","vbOriginalW","Gompertz1",
-                                  "Gompertz2","Gompertz3","Schnute"),
-                           x=NULL,y=NULL,data=NULL,max.len=500,max.wt=500) {
-  type <- match.arg(type)
-  if (class(x)=="formula") {
-    DF <- model.frame(x,data=data)
-    x <- DF[,2]
-    y <- DF[,1]
-    max.len <- max.wt <- max(y,na.rm=TRUE)
-  }
-  if ((!is.null(x) & is.null(y)) | (is.null(x) & !is.null(y))) stop("Both x (age) and y (length/weight) must be supplied if one is given.",call.=FALSE)
-  growthModelSim.default(type,x,y,max.len,max.wt)
-}
-
+#' @export growthModelSim
+#' 
 ## Main function
-growthModelSim.default <- function(type,x,y,max.len,max.wt) {
-  ## internal function to predict length/weight for plotting -- used in growSimPlot
-  predLength <- function(type,t,p1,p2,p3,p4) {
-    switch(type,
-           # p1=Linf, p2=K,  p3=to, p4 not used
-           vbTypical= {  sd <- p1*(1-exp(-p2*(t-p3))) },
-           # p1=Winf, p2=K,  p3=to, p4=b
-           vbTypicalW= { sd <- p1*(1-exp(-p2*(t-p3)))^p4 },         
-           # p1=Linf, p2=L0, p3=K,  p4 not used
-           vbOriginal= { sd <- (p1-(p1-p2)*exp(-p3*t)) },
-           # p1=Winf, p2=W0, p3=K,  p4=b
-           vbOriginalW= { sd <- (p1-(p1-p2)*exp(-p3*t))^p4 },
-           # p1=omega,p2=K,  p3=t0, p4 not used
-           vbGalucciQuinn= { sd <- (p1/p2)*(1-exp(-p2*(t-p3))) },
-           # p1=Linf, p2=L0, p3=ome,p4 not used
-           vbMooij= { sd <- p1-(p1-p2)*exp(-(p3/p1)*t) },
-           # p1=L1,   p2=L2, p3=K,  p4=b
-           vbSchnute= {
-             sd <- ((p1^p4)+((p2^p4)-(p1^p4))*((1-exp(-p3*(t-min(t))))/
-                                              (1-exp(-p3*(max(t)-min(t))))))^(1/p4)
-           },
-           # p1=Lo,   p2=G,  p3=g,  p4 not used
-           Gompertz1= { sd <- p1*(exp(p2*(1-exp(-p3*t)))) },
-           # p1=Linf, p2=g,  p3=t*, p4 not used
-           Gompertz2= { sd <- p1*(exp(-exp(-p2*(t-p3)))) },
-           # p1=Linf, p2=g,  p3=t0, p4 not used
-           Gompertz3= { sd <- p1*(exp((-1/p2)*exp(-p2*(t-p3)))) },
-           # p1=L1,   p2=L2, p3=c,  p4=d
-           Schnute= {
-             minage <- min(t)
-             maxage <- max(t)
-             diffage <- maxage-minage
-             if (p3==0) {
-               if (p4==0) {
-                 # Case 4
-                 sd <- p1*exp(log(p2/p1)*(t-minage)/diffage)
-               } else { 
-                 # Case 3
-                 sd <- (p1^p4+(p2^p4-p1^p4)*(t-minage)/diffage)^(1/p4)
-              }
-             } else {
-               if (p4==0) {
-                 # Case 2
-                 sd <- p1*exp(log(p2/p1)*(1-exp(-p3*(t-minage)))/(1-exp(-p3*diffage)))
-               } else { 
-                 # Case 1
-                 sd <- (p1^p4+(p2^p4-p1^p4)*((1-exp(-p3*(t-minage)))/
-                                            (1-exp(-p3*diffage))))^(1/p4) 
-               }
-             }
-           } # end Schnute
-    ) # end switch 
-    sd
-  } ## end predLength internal function
-  
-  ## internal function for constructing the plot
-  growthSimPlot <- function(type,x,y,max.y,t.max,p1,p2,p3,p4) {
-    t <- seq(0,t.max,length.out=t.max*20)
-    vals <- predLength(type,t,p1,p2,p3,p4)
-    ylbl <- ifelse (type %in% c("vbTypicalW","vbOriginalW"),"Weight","Length")
-    old.par <- par(mar=c(3.5,3.5,1,1), mgp=c(2,0.4,0), tcl=-0.2); on.exit(par(old.par))
-    if (is.null(x)||is.null(y)) plot(t,vals,xlab="Age",ylab=paste("Mean",ylbl),type="l",lwd=2,col="blue",ylim=c(0,max.y))
-    else {
-      plot(x,y,xlab="Age",ylab=ylbl,ylim=c(0,max.y),xlim=c(0,t.max))
-      lines(t,vals,type="l",lwd=2,col="blue")
-    }
-  } ## end growthSimPlot internal function
-  
+growthModelSim <- function(formula=NULL,data=NULL,
+                           type=c("vbTypical","vbOriginal","vbGallucciQuinn","vbMooij",
+                                  "vbSchnute","vbTypicalW","vbOriginalW",
+                                  "Gompertz1","Gompertz2","Gompertz3",
+                                  "Schnute"),
+                           max.len=500,max.wt=500) {
   ## Internal refresh function
   refresh <- function(...) {
     p1 <- relax::slider(no=1)
@@ -161,10 +87,31 @@ growthModelSim.default <- function(type,x,y,max.len,max.wt) {
       p4 <- 1
       t.max <- relax::slider(no=4)
     }
-  growthSimPlot(type,x,y,max.y,t.max,p1,p2,p3,p4)
-  } ## end internal refresh
+  iGrowthSimPlot(type,x,y,max.y,t.max,p1,p2,p3,p4)
+  } ## end Internal refresh
 
   ## begin main growthModelSim
+  # some checks and handle the formula
+  if (!is.null(formula)) {
+    # this allows the user to enter a type as the first argument as long as data is not given
+    if (is.character(formula) & is.null(data)) type <- formula
+    else {
+      tmp <- hndlFormula(formula,data,expNumR=1,expNumE=1)
+      if (!tmp$metExpNumR) stop("'vbStarts' must have only one LHS variable.",call.=FALSE)
+      if (!tmp$Rclass %in% c("numeric","integer")) stop("LHS variable must be numeric.",call.=FALSE)
+      if (!tmp$metExpNumE) stop("'vbStarts' must have only one RHS variable.",call.=FALSE)
+      if (!tmp$Eclass %in% c("numeric","integer")) stop("RHS variable must be numeric.",call.=FALSE)
+      # get the length and age vectors
+      y <- tmp$mf[,tmp$Rname[1]]
+      x <- tmp$mf[,tmp$Enames[1]]
+      # set maximum size to max observed data
+      max.len <- max.wt <- max(y,na.rm=TRUE)
+    }
+  } else {
+    # if no formula given then set x, y vectors to NULL
+    x <- y <- NULL
+  }
+  type <- match.arg(type)
   t.max <- ifelse(!is.null(x),max(x,na.rm=TRUE),20)
   max.y <- ifelse (type %in% c("vbTypicalW","vbOriginalW"),max.wt,max.len)
   if (!is.null(y)) max.y <- 1.1*max(y,na.rm=TRUE)
@@ -207,15 +154,15 @@ growthModelSim.default <- function(type,x,y,max.len,max.wt) {
              sl.defaults=c(  max.y,  0.3,     0,    1,     t.max),
              title = "Typical Von Bertalanffy Weight",pos.of.panel="left")
       }, # end vbTypical
-    vbGalucciQuinn= {
+    vbGallucciQuinn= {
       relax::gslider(refresh,prompt=TRUE,
              sl.names=   c("omega",  "K", "t_0", "Max Age"),
              sl.mins=    c(      0,    0,   -10,         5),
              sl.maxs=    c(    250,  0.5,    10,       100),
              sl.deltas=  c(      5, 0.01,   0.1,         1),
              sl.defaults=c(     75,  0.3,     0,     t.max),
-             title = "Galucci & Quinn Von Bertalanffy",pos.of.panel="left")
-      }, # end vbGalucciQuinn
+             title = "Gallucci & Quinn Von Bertalanffy",pos.of.panel="left")
+      }, # end vbGallucciQuinn
     vbMooij= {
       relax::gslider(refresh,prompt=TRUE,
              sl.names=   c("L_inf",    "L_0", "omega", "Max Age"),
@@ -272,3 +219,72 @@ growthModelSim.default <- function(type,x,y,max.len,max.wt) {
     } # end Schnute
   ) # end switch
 }
+
+
+
+
+## Internal function to predict length/weight for plotting -- used in iGrowthSimPlot
+iPredLength <- function(type,t,p1,p2,p3,p4) {
+  switch(type,
+         # p1=Linf, p2=K,  p3=to, p4 not used
+         vbTypical= {  sd <- p1*(1-exp(-p2*(t-p3))) },
+         # p1=Winf, p2=K,  p3=to, p4=b
+         vbTypicalW= { sd <- p1*(1-exp(-p2*(t-p3)))^p4 },         
+         # p1=Linf, p2=L0, p3=K,  p4 not used
+         vbOriginal= { sd <- (p1-(p1-p2)*exp(-p3*t)) },
+         # p1=Winf, p2=W0, p3=K,  p4=b
+         vbOriginalW= { sd <- (p1-(p1-p2)*exp(-p3*t))^p4 },
+         # p1=omega,p2=K,  p3=t0, p4 not used
+         vbGallucciQuinn= { sd <- (p1/p2)*(1-exp(-p2*(t-p3))) },
+         # p1=Linf, p2=L0, p3=ome,p4 not used
+         vbMooij= { sd <- p1-(p1-p2)*exp(-(p3/p1)*t) },
+         # p1=L1,   p2=L2, p3=K,  p4=b
+         vbSchnute= {
+           sd <- ((p1^p4)+((p2^p4)-(p1^p4))*((1-exp(-p3*(t-min(t))))/(1-exp(-p3*(max(t)-min(t))))))^(1/p4)
+         },
+         # p1=Lo,   p2=G,  p3=g,  p4 not used
+         Gompertz1= { sd <- p1*(exp(p2*(1-exp(-p3*t)))) },
+         # p1=Linf, p2=g,  p3=t*, p4 not used
+         Gompertz2= { sd <- p1*(exp(-exp(-p2*(t-p3)))) },
+         # p1=Linf, p2=g,  p3=t0, p4 not used
+         Gompertz3= { sd <- p1*(exp((-1/p2)*exp(-p2*(t-p3)))) },
+         # p1=L1,   p2=L2, p3=c,  p4=d
+         Schnute= {
+           minage <- min(t)
+           maxage <- max(t)
+           diffage <- maxage-minage
+           if (p3==0) {
+             if (p4==0) {
+               # Case 4
+               sd <- p1*exp(log(p2/p1)*(t-minage)/diffage)
+             } else { 
+               # Case 3
+               sd <- (p1^p4+(p2^p4-p1^p4)*(t-minage)/diffage)^(1/p4)
+             }
+           } else {
+             if (p4==0) {
+               # Case 2
+               sd <- p1*exp(log(p2/p1)*(1-exp(-p3*(t-minage)))/(1-exp(-p3*diffage)))
+             } else { 
+               # Case 1
+               sd <- (p1^p4+(p2^p4-p1^p4)*((1-exp(-p3*(t-minage)))/
+                                             (1-exp(-p3*diffage))))^(1/p4) 
+             }
+           }
+         } # end Schnute
+  ) # end switch 
+  sd
+} ## end iPredLength internal function
+
+## internal function for constructing the plot
+iGrowthSimPlot <- function(type,x,y,max.y,t.max,p1,p2,p3,p4) {
+  t <- seq(0,t.max,length.out=t.max*20)
+  vals <- iPredLength(type,t,p1,p2,p3,p4)
+  ylbl <- ifelse (type %in% c("vbTypicalW","vbOriginalW"),"Weight","Length")
+  old.par <- par(mar=c(3.5,3.5,1,1), mgp=c(2,0.4,0), tcl=-0.2); on.exit(par(old.par))
+  if (is.null(x)||is.null(y)) plot(t,vals,xlab="Age",ylab=paste("Mean",ylbl),type="l",lwd=2,col="blue",ylim=c(0,max.y))
+  else {
+    plot(x,y,xlab="Age",ylab=ylbl,ylim=c(0,max.y),xlim=c(0,t.max))
+    lines(t,vals,type="l",lwd=2,col="blue")
+  }
+} ## end iGrowthSimPlot internal function